-Refactored naming for Transfer actions and views -Fixed currencies and cards not loading into form if modelstate is not valid -Fixed validations not working in WalletTransactionController

+++ VirtualWallet/VirtualWallet.DTO/ViewModels/TransferViewModels/TransferViewModel.cs
@@ -13,7 +13,7 @@ namespace VirtualWallet.Dto.ViewModels.TransferViewModels
         public bool HasCardSender { get; set; }
 
         [Required(ErrorMessage = "Please enter {0}!")]
-        [Range(0.01, double.MaxValue, ErrorMessage = "The {0} must be greater than 0.")]
+        [Range(0.01, (double)decimal.MaxValue, ErrorMessage = "The {0} must be greater than 0.")]
         public decimal Amount { get; set; }
 
         [Required(ErrorMessage = "Please select a card!")]
